Utils cleanup. Module tests.utils should only be used for tests.
